+++ changelog.md
@@ -6,6 +6,11 @@
 
 **Fixes**
  * Add support for multiple classloaders when using CoerceUtils.
+ * Issue#691
+ * Issue#644
+
+**Features**
+ * Both JPA Field (new) and Property (4.2.6 and earlier) Access are now supported.
 
 ## 4.2.6
 **Fixes**